Migrate to vite Fixed code that causes errors
