+++ 07 Debugging React Apps/01-starting-project/src/components/CourseGoals/CourseGoalItem/CourseGoalItem.jsx
